Replace Metrics with Standard exporter
* Move metrics display responsibility to Standard exporter
